Dl4j 1.0.0 beta3

+++ src/main/java/ampcontrol/model/training/model/vertex/ChannelMultVertex.java
@@ -35,7 +35,7 @@ public class ChannelMultVertex extends GraphVertex {
     }
 
     @Override
-    public int numParams(boolean backprop) {
+    public long numParams(boolean backprop) {
         return 0;
     }
