+++ src/test/java/ampcontrol/model/training/model/evolve/fitness/FitnessPolicyTrainingTest.java
@@ -2,10 +2,10 @@ package ampcontrol.model.training.model.evolve.fitness;
 
 import ampcontrol.model.training.listen.MockModel;
 import ampcontrol.model.training.model.ModelAdapter;
-import org.deeplearning4j.eval.IEvaluation;
 import org.deeplearning4j.nn.api.Model;
 import org.deeplearning4j.optimize.api.TrainingListener;
 import org.junit.Test;
+import org.nd4j.evaluation.IEvaluation;
 import org.nd4j.linalg.dataset.api.iterator.DataSetIterator;
 
 import java.util.ArrayList;
@@ -42,8 +42,8 @@ public class FitnessPolicyTrainingTest {
             }
 
             @Override
-            public int numParams() {
-                return (int) (5* 1e4);
+            public long numParams() {
+                return (long)(5* 1e4);
             }
 
             @Override
